+++ src/detection/terminalshell/terminalshell_windows.c
@@ -143,7 +143,7 @@ static void setShellInfoDetails(FFShellResult* result)
         ffStrbufSetS(&result->prettyName, "Windows PowerShell ISE");
     else if(ffStrbufIgnCaseEqualS(&result->prettyName, "cmd"))
     {
-        ffStrbufClear(&result->prettyName);
+        ffStrbufSetS(&result->prettyName, "CMD");
 
         FF_AUTO_CLOSE_FD HANDLE snapshot = NULL;
         while(!(snapshot = CreateToolhelp32Snapshot(TH32CS_SNAPMODULE, result->pid)) && GetLastError() == ERROR_BAD_LENGTH) {}
@@ -156,15 +156,13 @@ static void setShellInfoDetails(FFShellResult* result)
             {
                 if(wcsncmp(module.szModule, L"clink_dll_", strlen("clink_dll_")) == 0)
                 {
-                    ffStrbufAppendS(&result->prettyName, "CMD (with Clink ");
+                    ffStrbufAppendS(&result->prettyName, " (with Clink ");
                     getProductVersion(module.szExePath, &result->prettyName);
                     ffStrbufAppendC(&result->prettyName, ')');
                     break;
                 }
             }
         }
-        if(result->prettyName.length == 0)
-            ffStrbufAppendS(&result->prettyName, "Command Prompt");
     }
     else if(ffStrbufIgnCaseEqualS(&result->prettyName, "nu"))
         ffStrbufSetS(&result->prettyName, "nushell");
@@ -345,7 +343,7 @@ static void setTerminalInfoDetails(FFTerminalResult* result)
             : "Windows Terminal"
         );
     else if(ffStrbufIgnCaseEqualS(&result->prettyName, "conhost"))
-        ffStrbufSetStatic(&result->prettyName, "Console Window Host");
+        ffStrbufSetStatic(&result->prettyName, "Windows Console");
     else if(ffStrbufIgnCaseEqualS(&result->prettyName, "Code"))
         ffStrbufSetStatic(&result->prettyName, "Visual Studio Code");
     else if(ffStrbufIgnCaseEqualS(&result->prettyName, "explorer"))